+++ authority-portal-frontend/src/app/pages/control-center-organization-edit-page/state/control-center-organization-edit-page-state-impl.ts
@@ -74,7 +74,7 @@ export class ControlCenterOrganizationEditPageStateImpl {
   ): HeaderBarConfig {
     return {
       title: organization.name,
-      subtitle: 'Edit your organization details',
+      subtitle: 'Edit Your Organization Profile',
       headerActions: [],
     };
   }