+++ authority-portal-frontend/src/app/pages/control-center-user-edit-page/control-center-user-edit-page/control-center-user-edit-page.form-model.ts
@@ -1,0 +1,7 @@
+import {ɵFormGroupRawValue} from '@angular/forms';
+import {UserEditFormModel} from '../../../shared/components/business/user-edit-form/user-edit-form-model';
+
+export interface ControlCenterUserEditPageFormModel extends UserEditFormModel {}
+
+export type ControlCenterUserEditPageFormValue =
+  ɵFormGroupRawValue<ControlCenterUserEditPageFormModel>;