+++ authority-portal-frontend/src/app/shared/components/business/user-edit-form/user-edit-form-model.ts
@@ -1,0 +1,11 @@
+import {FormControl, ɵFormGroupRawValue} from '@angular/forms';
+
+export interface UserEditFormModel {
+  firstName: FormControl<string>;
+  lastName: FormControl<string>;
+  email: FormControl<string>;
+  jobTitle: FormControl<string>;
+  phoneNumber: FormControl<string>;
+}
+
+export type UserEditFormValue = ɵFormGroupRawValue<UserEditFormModel>;