+++ authority-portal-frontend/src/app/shared/components/business/user-onboard-form/user-onboard-form-model.ts
@@ -1,0 +1,10 @@
+import {FormControl, ɵFormGroupRawValue} from '@angular/forms';
+
+export interface UserOnboardFormModel {
+  firstName: FormControl<string>;
+  lastName: FormControl<string>;
+  jobTitle: FormControl<string>;
+  phoneNumber: FormControl<string>;
+}
+
+export type UserOnboardFormValue = ɵFormGroupRawValue<UserOnboardFormModel>;