+++ CHANGELOG.md
@@ -14,6 +14,9 @@ please see [changelog_updates.md](docs/dev/changelog_updates.md).
 #### Minor
 
 - Control Center: Edit Own Organization Details Page
+- Control Center: Edit Own User Details Page
+- Improved User Profiles slightly by reordering items
+- Improved Organization Profiles slightly by reordering items
 
 #### Patch